refactor: make Index system ITensors.jl-compatible (#196)
* refactor: make Index system ITensors.jl-compatible (u64 ID, id+tags equality)

Align the index system with ITensors.jl (non-QN version) for HDF5 format
compatibility. Three key changes:

1. DynId: u128 → u64 (matches ITensors.jl's IDType = UInt64)
2. Equality/Hash: id-only → id+tags (matches ITensors.jl's id+plev+tags
   with plev conceptually fixed at 0)
3. Hash implementations added to SmallString and TagSet

Updates across all layers: core Rust, C API, Julia bindings, Python
bindings, tests, and documentation.


* fix: remove unused Index import in direct_sum.rs


* fix: remove Index usage from direct_sum test code

Replace Index::new() calls with clone() in tests to match the removal
of the Index import from non-test code.


---------

+++ crates/tensor4all-core/src/defaults/contract.rs
@@ -762,7 +762,7 @@ mod tests {
     use crate::defaults::Index;
     use num_complex::Complex64;
 
-    fn make_test_tensor(shape: &[usize], ids: &[u128]) -> TensorDynLen {
+    fn make_test_tensor(shape: &[usize], ids: &[u64]) -> TensorDynLen {
         let indices: Vec<DynIndex> = ids
             .iter()
             .zip(shape.iter())
@@ -1232,7 +1232,7 @@ mod tests {
     // contract_connected tests
     // ========================================================================
 
-    fn make_dense_tensor(shape: &[usize], ids: &[u128]) -> TensorDynLen {
+    fn make_dense_tensor(shape: &[usize], ids: &[u64]) -> TensorDynLen {
         let indices: Vec<DynIndex> = ids
             .iter()
             .zip(shape.iter())
